refactor(manager): move some folder to manager
