+++ manager/delivery/httpserver/userhandler/user_register.go
